Support PTC motor thermistor (#46) * Auto detect and support 1k PTC thermistor present in some BBSHD motors for motor temperature reading. * Use 10bit ADC resolution for temperature readings. * Fix negative temperatures did not print correctly in eventlog.

+++ src/firmware/adc.h
@@ -15,7 +15,7 @@ void adc_init();
 void adc_process();
 
 uint8_t adc_get_throttle();
-uint8_t adc_get_temperature_contr();
-uint8_t adc_get_temperature_motor();
+uint16_t adc_get_temperature_contr();
+uint16_t adc_get_temperature_motor();
 
 #endif
